split batch/task codes to tasks.py
