+++ requirements.txt
@@ -8,3 +8,6 @@ opengraph==0.5
 mock==2.0.0
 ua-parser==0.7.3
 user-agents==1.0.1
+google-api-python-client==1.6.4
+google-auth==1.2.1
+google-auth-httplib2==0.0.3
